Data.TSTP

Documentation

data F

Main formula type, it contains all the elements and information of a TSTP formula definition. While name, role, and formula are self-explanatory, source is a messy meta-language in itself, different ATPs may embed different amounts of information in it.

Constructors

F

Fields

name :: String

role :: Role

formula :: Formula

source :: Source

data Role

Formula roles.

Constructors

Axiom

Hypothesis

Definition

Assumption

Lemma

Theorem

Conjecture

NegatedConjecture

Plain

FiDomain

FiFunctors

FiPredicates

Type

Unknown

Formulas and terms

data Formula

first-order logic formula.

Constructors

BinOp Formula BinOp Formula
Binary connective application

InfixPred Term InfixPred Term
Infix predicate application

PredApp AtomicWord [Term]
Predicate application

Quant Quant [V] Formula
Quantified Formula

(:~:) Formula
Negation

data Term

First-order logic terms.

Constructors

Var V
Variable

NumberLitTerm Rational
Number literal

DistinctObjectTerm String
Double-quoted item

FunApp AtomicWord [Term]
Function symbol application (constants are encoded as nullary functions)

Show instances

Formula, Term and other data types in this section have Show instances that allow pretty-printing of Formulas and Show [Formula] is an especial instance that print its contents as sequence of implications

>>> let f1 = PredApp (AtomicWord "a") []
>>> let f2 = PredApp (AtomicWord "b") []
>>> let f3 = (BinOp (PredApp (AtomicWord "a") []) (:&:) (PredApp (AtomicWord "b") []))
>>> f1
a
>>> f2
b
>>> f3
a ∧ b
>>> [f1,f2,f3]
{ a b : Set} → a → b → a ∧ b

Some syntax sugar is also present

>>> PredApp (AtomicWord "$false") []
⊥

newtype V

Variable

Constructors

V String

data BinOp

Binary formula connectives.

Constructors

(:<=>:)
↔ Equivalence

(:=>:)
→ Implication

(:<=:)
← Reverse Implication

(:&:)
∧ AND

(:|:)
∨ OR

(:~&:)
⊼ NAND

(:~|:)
⊽ NOR

(:<~>:)
⊕ XOR

data InfixPred

Infix connectives of the form Term -> Term -> Formula.

Constructors

(:=:)
=

(:!=:)
≠

data Quant

Quantifier specification.

Constructors

All
∀

Exists
∃

newtype AtomicWord

Constructors

AtomicWord String

Source information

data Source

Source main purpose is to provide all the information regarding the deductive process that lead to a given formula. Information about the rules applied along with parent formulas and SZS status are among the information you might expect from this field.

Constructors

Source String

Inference Rule [Info] [Parent]

Introduced IntroType [Info]

File String (Maybe String)

Theory Theory [Info]

Creator String [Info]

Name String

NoSource

data Rule

Deduction rule applied.

Constructors

Simplify

Negate

Canonicalize

Strip

NewRule String

data Parent

Parent formula information.

Constructors

Parent String [GTerm]

Functions

isBottom :: F -> Bool

isBottom f, test whether f = ⊥.

bottom :: Formula

bottom = ⊥.

freeVarsF :: Formula -> Set V

freeVarsF f, returns a Set of all free variables of f.

freeVarsT :: Term -> Set V

freeVarsT t, returns a Set of all free variables of t.

getFreeVars :: [Formula] -> [V]

getFreeVars f, given a list of formulas f return all free variables in the formulas.

Unused types

The following types are required to have full support of the TSTP syntax but haven't been used yet in tstp2agda aside from the parser.

data IntroType

Constructors

Definition_

AxiomOfChoice

Tautology

Assumption_

UnknownType

data Theory

Constructors

Equality

AC

data Info

Constructors

Description String

IQuote String

Status Status

Function String [GTerm]

InferenceInfo Rule String [GTerm]

AssumptionR [String]

Refutation Source

data Status

Constructors

Suc

Unp

Sap

Esa

Sat

Fsa

Thm

Eqv

Tac

Wec

Eth

Tau

Wtc

Wth

Cax

Sca

Tca

Wca

Cup

Csp

Ecs

Csa

Cth

Ceq

Unc

Wcc

Ect

Fun

Uns

Wuc

Wct

Scc

Uca

Noc

Unk

data GData

Metadata (the general_data rule in TPTP's grammar)

Constructors

GWord AtomicWord

GApp AtomicWord [GTerm]

GVar V

GNumber Rational

GDistinctObject String

GFormulaData String Formula

GFormulaTerm String Term

data GTerm

Metadata (the general_term rule in TPTP's grammar)

Constructors

ColonSep GData GTerm

GTerm GData

GList [GTerm]